Added Omega* integrals from table F-1b of Ivchenko et al.
ChangeLog documents all changes.

Overall, I had to create a new variable to store the table, and an additional object.
Along the way, I also cleaned up some stuff and refactored other.
Finally, in the user package, I added code to plot the Omega* integrals.  This
code verifies that the tables are read correctly, and also that the interpoations
are working.